% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarkUniformity.R
\name{benchmarkUniformity}
\alias{benchmarkUniformity}
\title{Runs simulation to confirm uniformity of residuals under H0}
\usage{
benchmarkUniformity(dataModelCreator, nSim = 100, plot = T)
}
\arguments{
\item{plot}{should a plot be created}

\item{dataCreator}{a function that returns a list with two elements, data and model. See help for details}

\item{nsim}{number of simulations}
}
\description{
Runs simulation to confirm uniformity of residuals under H0
}
\examples{
library(lme4)

dataModelCreator <- function(){
  data = createData(sampleSize = 100, overdispersion = 0, family = poisson())
  model <- glmer(observedResponse ~ Environment1 + (1|group) + (1|ID), data = data, family = "poisson")  
  return(list(data=data, model = model))
}

benchmarkUniformity(dataModelCreator = dataModelCreator)
}

