% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{checkDots}
\alias{checkDots}
\title{Check dot operator}
\usage{
checkDots(name, value, ...)
}
\description{
Check dot operator
}
\details{
modified from https://github.com/lcolladotor/dots
}
\keyword{internal}
