% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateResiduals.R
\name{checkModel}
\alias{checkModel}
\title{Check if the fitted model is supported by DHARMa}
\usage{
checkModel(fittedModel, stop = F)
}
\arguments{
\item{fittedModel}{a fitted model}

\item{stop}{whether to throw an error if the model is not supported by DHARMa}
}
\description{
The function checks if the fitted model is supported by DHARMa, and if there are other issues that could create problems
}
\details{
The main purpose of this function os to check if the fitted model class is supported by DHARMa. The function additionally checks for properties of the fitted model that could create problems for calculating residuals or working with the resuls in DHARMa.
}
\examples{

testData = createData(sampleSize = 200, overdispersion = 0.5, randomEffectVariance = 0.5, family = gaussian(), hasNA = T)
fittedModel <- lm(observedResponse ~ Environment1 , data = testData)
res <- simulateResiduals(fittedModel) # throws NA message

# get the indices of the rows that were used to fit the model
sel = as.numeric(rownames(model.frame(fittedModel)))

# now use the indices when plotting against other variables
plotResiduals(res, form = testData$Environment1[sel])
}
\keyword{internal}
