% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateResiduals.R
\name{checkSimulations}
\alias{checkSimulations}
\title{Check simulated data}
\usage{
checkSimulations(simulatedResponse, nObs, nSim, bigData = FALSE)
}
\arguments{
\item{simulatedResponse}{the simulated response}

\item{nObs}{number of observations}

\item{nSim}{number of simulations}
}
\description{
The function checks if the simulated data seems fine
}
\keyword{internal}
