% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createData.R
\name{createBinomialData}
\alias{createBinomialData}
\title{creates Poisson data overdispersion and random intercept}
\usage{
createBinomialData(replicates = 1, sampleSize = 2000, intercept = 0,
  slope = 1, numGroups = 10, randomEffectVariance = 1,
  overdispersion = 0)
}
\description{
creates Poisson data overdispersion and random intercept
}

