% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DHARMa.R
\name{createDHARMa}
\alias{createDHARMa}
\title{Convert simulated residuals to a DHARMa object}
\usage{
createDHARMa(scaledResiduals = NULL, simulatedResponse = NULL,
  observedResponse = NULL, fittedPredictedResponse = NULL,
  integerResponse = F)
}
\arguments{
\item{scaledResiduals}{scaled residuals from a simulation, e.g. Bayesian p-values}

\item{simulatedResponse}{matrix with rows being observations and colums being simulations}

\item{fittedPredictedResponse}{fitted predicted response. Optional, but will be neccessary for some plots. If scaled residuals are Bayesian p-values, using the median posterior prediction as fittedPredictedResponse is recommended.}
}
\description{
Convert simulated residuals to a DHARMa object
}
\details{
The use of this function is to convert simulated residuals (e.g. from a point estimate, or Bayesian p-values) to a DHARMa object, to make use of the plotting / test functions in DHARMa
}
\note{
either scaled residuals or (simulatedResponse AND observed response) have to be provided
}

