% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createData.R
\name{createPoissonData}
\alias{createPoissonData}
\title{creates Poisson data overdispersion and random intercept}
\usage{
createPoissonData(replicates = 1, sampleSize = 2000, intercept = 0,
  slope = 1, numGroups = 10, randomEffectVariance = 1,
  overdispersion = 0.5)
}
\description{
creates Poisson data overdispersion and random intercept
}

