% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{DHARMa.ecdf}
\alias{DHARMa.ecdf}
\title{Modified ECDF function.}
\usage{
DHARMa.ecdf(x)
}
\description{
Modified ECDF function.
}
\details{
ensures symmetric ECDF (standard ECDF is <), and that 0 / 1 values are only produced if the data is strictly < > than the observed data.
}
\keyword{internal}
