% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DHARMa.R
\docType{package}
\name{DHARMa}
\alias{DHARMa}
\alias{DHARMa-package}
\title{DHARMa - Residual Diagnostics for HierArchical (Multi-level / Mixed) Regression Models}
\description{
The DHARMa package uses a simulation-based approach to create readily interpretable scaled residuals from fitted generalized linear mixed models. Currently supported are all 'merMod' classes from 'lme4' ('lmerMod', 'glmerMod'), 'glm' (including 'negbin' from 'MASS', but excluding quasi-distributions) and 'lm' model classes. Alternatively, externally created simulations, e.g. posterior predictive simulations from Bayesian software such as 'JAGS', 'STAN', or 'BUGS' can be processed as well. The resulting residuals are standardized to values between 0 and 1 and can be interpreted as intuitively as residuals from a linear regression. The package also provides a number of plot and test functions for typical model mispecification problem, such as over/underdispersion, zero-inflation, and spatial and temporal autocorrelation.
}
\details{
See index / vignette for details
}
\examples{
vignette("DHARMa", package="DHARMa")
}
\seealso{
\code{\link{simulateResiduals}}
}

