% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesisTestsOverdispersion.R
\name{dispersionTest}
\alias{dispersionTest}
\title{A wrapper for a number of parametric overdispersion tests}
\usage{
dispersionTest(model, type = 1)
}
\arguments{
\item{model}{a fitted model}

\item{type}{the test type. See details.}
}
\description{
A wrapper for a number of parametric overdispersion tests
}
\details{
This function is a wrapper for a number of parametric overdispersion tests. I am not actively developing these tests for the DHARMa package. The main purpose of providing these functions here is benchmarking agains the simulation-based tests in DHARMa. That being said, feel 

Type 1 is from Harrison et al. 2014 (Using observation-level random effects to model overdispersion in count data in ecology and evolution), but also include the function to test for overdispersion with Boostraping. 

Type 2 is the function dispersion_glmer in the blmeco package. From the help there: "This function has been posted on the R-helplist. It seems to have been written or motivated by D. Bates. Here is the URL, where we downloaded the function: https://stat.ethz.ch/pipermail/r-sig-mixed-models/2011q1/015392.html"
}

