% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DHARMa.R
\name{ensureDHARMa}
\alias{ensureDHARMa}
\title{Ensures that an object is of class DHARMa}
\usage{
ensureDHARMa(simulationOutput, convert = F)
}
\arguments{
\item{simulationOutput}{a DHARMa simulation output or an object that can be converted into a DHARMa simulation output}

\item{convert}{if TRUE, attempts to convert model + numeric to DHARMa, if "Model", converts only supported models to DHARMa}
}
\value{
an object of class DHARMa
}
\description{
Ensures that an object is of class DHARMa
}
\details{
The
}
\keyword{internal}
