% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DHARMa.R
\name{ensurePredictor}
\alias{ensurePredictor}
\title{Ensures that an object is of class DHARMa}
\usage{
ensurePredictor(simulationOutput, predictor = NULL)
}
\arguments{
\item{simulationOutput}{a DHARMa simulation output or an object that can be converted into a DHARMa simulation output}

\item{predictor}{a predictor}
}
\description{
Ensures that an object is of class DHARMa
}
\details{
The
}
\keyword{internal}
