% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compatibility.R
\name{fitted.gam}
\alias{fitted.gam}
\title{This function overwrites the standard fitted function for GAM}
\usage{
\method{fitted}{gam}(x)
}
\arguments{
\item{x}{fitted model}
}
\description{
This function overwrites the standard fitted function for GAM
}
\note{
See explanation at
}

