% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compatibility.R
\name{getFamily}
\alias{getFamily}
\alias{getFamily.default}
\alias{getFamily.phyloglm}
\title{Get model family}
\usage{
getFamily(object, ...)

\method{getFamily}{default}(object, ...)

\method{getFamily}{phyloglm}(object, ...)
}
\arguments{
\item{object}{a fitted model}

\item{...}{additional parameters to be passed on}
}
\description{
Wrapper to get the family of a fitted model

checks if the fitted model excluded NA values
}
\details{
Checks if the fitted model excluded NA values
}
\seealso{
\code{\link{getObservedResponse}}, \code{\link{getSimulations}}, \code{\link{getRefit}}, \code{\link{getFixedEffects}}, \code{\link{getFitted}}
}
\author{
Florian Hartig
}
