% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compatibility.R
\name{getFitted.gam}
\alias{getFitted.gam}
\title{This function overwrites the standard fitted function for GAM}
\usage{
\method{getFitted}{gam}(object, ...)
}
\arguments{
\item{object}{fitted model}

\item{...}{arguments to be passed on to stats::fitted}
}
\description{
This function overwrites the standard fitted function for GAM
}
\note{
See explanation at
}
