% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{getFormulaPredictors}
\alias{getFormulaPredictors}
\title{Get predictors specified in a formula.}
\usage{
getFormulaPredictors(simulationOutput, formula)
}
\arguments{
\item{simulationOutput}{simulationOutput (DHARMa object).}

\item{formula}{the formula to be evaluated, typically consisting of one variable except for plotResiduals (optional multiple variables) and testSpatialAutocorrelation (mandatory x and y).}
}
\description{
Gets additional predictor(s) specified as a formula for plotResiduals, recalculateResiduals, testCategorical, testQuantiles, testTemporalAutocorrelation, testSpatialAutocorrelation. Predictors are extracted from the data frame stored in the model object. If observations were removed during model fitting due to NAs, these rows are excluded automatically by getFormulaPredictors.
}
\keyword{internal}
