% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compatibility.R
\name{getPredictions}
\alias{getPredictions}
\alias{getPredictions.default}
\alias{getPredictions.HLfit}
\alias{getPredictions.MixMod}
\title{Get model predictions}
\usage{
getPredictions(object, ...)

\method{getPredictions}{default}(object, type = "response",
  re.form = ~0, ...)

\method{getPredictions}{HLfit}(object, type = "response", re.form = ~0,
  ...)

\method{getPredictions}{MixMod}(object, type_pred = "response",
  type = "mean_subject", ...)
}
\arguments{
\item{object}{a fitted model}

\item{...}{additional parameters}
}
\description{
Extract the predictions of a fitted model
}
\details{
The purpose of this function is to savely extract the predictions of the fitted model classes
}
\author{
Florian Hartig
}
