% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compatibility.R
\name{getPredictorNames}
\alias{getPredictorNames}
\alias{getPredictorNames.default}
\title{Get predictor names}
\usage{
getPredictorNames(object, ...)

\method{getPredictorNames}{default}(object, ...)
}
\arguments{
\item{object}{a fitted model.}

\item{...}{additional parameters to be passed on.}
}
\description{
Wrapper to get the names of the predictors used in the model
}
\seealso{
\link{getObservedResponse}, \link{getSimulations}, \link{getRefit}, \link{getFixedEffects}, \link{getFitted}
}
\author{
Florian Hartig
}
