% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{getQuantile}
\alias{getQuantile}
\title{Quantile calculations}
\usage{
getQuantile(simulations, observed, n, nSim, integerResponse)
}
\arguments{
\item{observed}{vector with observed data}

\item{simlations}{simulations}
}
\description{
Quantile calculations
}
\keyword{internal}
