% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compatibility.R
\name{hasNA}
\alias{hasNA}
\title{has NA}
\usage{
hasNA(object)
}
\arguments{
\item{object}{a fitted model}
}
\description{
checks if the fitted model excluded NA values
}
\details{
Checks if the fitted model excluded NA values
}
