% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{hasWeights}
\alias{hasWeights}
\title{Check if a model has (problematic) prior weights}
\usage{
hasWeights(fittedModel)
}
\arguments{
\item{fittedModel}{a fitted model}
}
\description{
For n/k models, weights can either represent the number of trials or actual prior weights on the likelihood. This function distinguishes the two cases and returns TRUE if the model has prior weights that are not the number of trials.
}
\keyword{internal}
