% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compatibility.R
\name{hasWeigths}
\alias{hasWeigths}
\title{Check weights}
\usage{
hasWeigths(object, ...)
}
\arguments{
\item{object}{a fitted model}

\item{...}{additional parameters}
}
\description{
Checks if a model was fit with the weight argument
}
\details{
The purpose of this function is to check if a model was fit with the weights
}
\author{
Florian Hartig
}
