% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{hist.DHARMa}
\alias{hist.DHARMa}
\title{Histogram of DHARMa residuals}
\usage{
\method{hist}{DHARMa}(x, breaks = seq(-0.02, 1.02, len = 53),
  col = c(.Options$DHARMaSignalColor, rep("lightgrey", 50),
  .Options$DHARMaSignalColor), main = "Hist of DHARMa residuals",
  xlab = "Residuals (outliers are marked red)", cex.main = 1, ...)
}
\arguments{
\item{x}{a DHARMa simulation output (class DHARMa).}

\item{breaks}{breaks for hist() function.}

\item{col}{color for histogram bars.}

\item{main}{plot title.}

\item{xlab}{plot x-axis label.}

\item{cex.main}{plot cex.main.}

\item{...}{other arguments to be passed on to hist().}
}
\description{
The function produces a histogram from a DHARMa output. Outliers are marked in red.
}
\details{
The function calls hist() to create a histogram of the scaled residuals. Outliers are marked red as default but it can be changed by setting \code{options(DHARMaSignalColor = "red")} to a different color. See \code{getOption("DHARMaSignalColor")} for the current setting.
}
\examples{
testData = createData(sampleSize = 200, family = poisson(),
                      fixedEffects = c(1,1),
                      randomEffectVariance = 1, numGroups = 10)
testData$Environment2[1] = NA
fittedModel <- glm(observedResponse ~ Environment1 + Environment2,
                   family = "poisson", data = testData)
simulationOutput <- simulateResiduals(fittedModel = fittedModel)

######### main plotting function #############

# for all functions, quantreg = T will be more
# informative, but slower

plot(simulationOutput, quantreg = FALSE)

#############  Distribution  ######################

plotQQunif(simulationOutput = simulationOutput,
           testDispersion = FALSE,
           testUniformity = FALSE,
           testOutliers = FALSE)

hist(simulationOutput )




}
\seealso{
\link{plotSimulatedResiduals}, \link{plotResiduals}
}
