% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DHARMa.R
\name{plot.DHARMa}
\alias{plot.DHARMa}
\title{Plot simulated residuals}
\usage{
\method{plot}{DHARMa}(simulationOutput, ...)
}
\arguments{
\item{simulationOutput}{an object with simulated residuals created by \code{\link{simulateResiduals}}}

\item{...}{options of \code{\link{plotSimulatedResiduals}}}
}
\description{
Plot simulated residuals
}

