% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runBenchmarks.R
\name{plot.DHARMaBenchmark}
\alias{plot.DHARMaBenchmark}
\title{Plots DHARMa benchmarks}
\usage{
\method{plot}{DHARMaBenchmark}(x, ...)
}
\arguments{
\item{x}{Object of class DHARMaBenchmark, created by \link{runBenchmarks}.}

\item{...}{Parameters to pass to the plot function.}
}
\description{
The function plots the result of an object of class DHARMaBenchmark, created by \link{runBenchmarks}.
}
\details{
The function will create two types of plots, depending on whether the run contains only a single value (or no value) of the control parameter, or whether a vector of control values is provided:

If a single or no value of the control parameter is provided, the function will create boxplots of the estimated p-values, with the number of significant p-values plotted to the left.

If a control parameter is provided, the function will plot the proportion of significant p-values against the control parameter, with 95\% CIs based on the performed replicates displayed as confidence bands.
}
\seealso{
\link{runBenchmarks}
}
