% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotResiduals.R
\name{plotConventionalResiduals}
\alias{plotConventionalResiduals}
\title{Conventional residual plot}
\usage{
plotConventionalResiduals(fittedModel)
}
\arguments{
\item{fittedModel}{a fitted model object}
}
\description{
Convenience function to draw conventional residual plots
}
