% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotQQunif}
\alias{plotQQunif}
\title{Quantile-quantile plot for a uniform distribution}
\usage{
plotQQunif(simulationOutput, testUniformity = TRUE, testOutliers = TRUE,
  testDispersion = TRUE, ...)
}
\arguments{
\item{simulationOutput}{a DHARMa simulation output (class DHARMa).}

\item{testUniformity}{if T, the function \link{testUniformity} will be called and the result will be added to the plot.}

\item{testOutliers}{if T, the function \link{testOutliers} will be called and the result will be added to the plot.}

\item{testDispersion}{if T, the function \link{testDispersion} will be called and the result will be added to the plot.}

\item{...}{arguments to be passed on to \link[gap:qqunif]{gap::qqunif}.}
}
\description{
The function produces a uniform quantile-quantile plot from a DHARMa output. Optionally, tests for uniformity, outliers and dispersion can be added.
}
\details{
The function calls qqunif() from the R package gap to create a quantile-quantile plot for a uniform distribution, and overlays tests for particular distributional problems as specified.
When tests are displayed, significant p-values are highlighted in the color red by default. This can be changed by setting \code{options(DHARMaSignalColor = "red")} to a different color. See \code{getOption("DHARMaSignalColor")} for the current setting.
}
\examples{
testData = createData(sampleSize = 200, family = poisson(),
                      fixedEffects = c(1,1),
                      randomEffectVariance = 1, numGroups = 10)
testData$Environment2[1] = NA
fittedModel <- glm(observedResponse ~ Environment1 + Environment2,
                   family = "poisson", data = testData)
simulationOutput <- simulateResiduals(fittedModel = fittedModel)

######### main plotting function #############

# for all functions, quantreg = T will be more
# informative, but slower

plot(simulationOutput, quantreg = FALSE)

#############  Distribution  ######################

plotQQunif(simulationOutput = simulationOutput,
           testDispersion = FALSE,
           testUniformity = FALSE,
           testOutliers = FALSE)

hist(simulationOutput )




}
\seealso{
\link{plotSimulatedResiduals}, \link{plotResiduals}
}
