% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSimulatedResiduals.R
\name{plotResiduals}
\alias{plotResiduals}
\title{Generic residual plot with either spline or quantile regression}
\usage{
plotResiduals(pred, res, quantreg = T, ...)
}
\arguments{
\item{pred}{predictor variable}

\item{res}{result variable}

\item{quantreg}{should a quantile regression be performed. If F, a smooth spline will be}
}
\description{
Generic residual plot with either spline or quantile regression
}

