% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotResidualsAll}
\alias{plotResidualsAll}
\title{Plot residuals against all predictors of the model.}
\usage{
plotResidualsAll(simulationOutput, form = NULL, quantreg = NULL,
  rank = TRUE, asFactor = NULL, smoothScatter = NULL,
  quantiles = c(0.25, 0.5, 0.75), absoluteDeviation = FALSE, ...)
}
\arguments{
\item{simulationOutput}{an object with simulated residuals created by \link{simulateResiduals}. For further parameters see plotResiduals().}
}
\description{
Internal function to plot residuals against all predictors of the model.
}
