% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSimulatedResiduals.R
\name{plotSimulatedResiduals}
\alias{plotSimulatedResiduals}
\title{creates Poisson data overdispersion and random intercept}
\usage{
plotSimulatedResiduals(simulationOutput)
}
\description{
creates Poisson data overdispersion and random intercept
}

