% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DHARMa.R
\name{print.DHARMa}
\alias{print.DHARMa}
\title{Print simulated residuals}
\usage{
\method{print}{DHARMa}(x, ...)
}
\arguments{
\item{x}{An object with simulated residuals created by \link{simulateResiduals}.}

\item{...}{Optional arguments for compatibility with the generic function, no function implemented.}
}
\description{
Print simulated residuals
}
