% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesisTestsPerformanceSimulation.R
\name{simulatePvaluesDispersion}
\alias{simulatePvaluesDispersion}
\title{simulates null}
\usage{
simulatePvaluesDispersion(overdispersion, n = 20, alpha = 0.05)
}
\value{
list of various named objects. Within each, order is dispersion_glmer, overdisp_fun, DHARMa
}
\description{
simulates null
}

