% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateResiduals.R
\name{simulateResiduals}
\alias{simulateResiduals}
\title{Creates scaled residuals by simulation}
\usage{
simulateResiduals(fittedModel, n = 250, refit = F)
}
\arguments{
\item{fittedModel}{fitted model, currently restricted to lme4 models}
}
\description{
Creates scaled residuals by simulation
}

