% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{testGeneric}
\alias{testGeneric}
\title{Generic simulation test of a summary statistic}
\usage{
testGeneric(simulationOutput, summary, alternative = c("greater", "two.sided",
  "less"), plot = T, methodName = "DHARMa generic simulation test")
}
\description{
Generic simulation test of a summary statistic
}
\seealso{
\code{\link{testResiduals}}, \code{\link{testUniformity}}, \code{\link{testDispersion}}, \code{\link{testZeroInflation}}, \code{\link{testTemporalAutocorrelation}}, \code{\link{testSpatialAutocorrelation}}
}
\author{
Florian Hartig
}
