% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compatibility.R
\name{testModel}
\alias{testModel}
\title{Test DHARMa compatibility}
\usage{
testModel(fittedModel)
}
\arguments{
\item{fittedModel}{the fitted model}
}
\description{
This helper function tests the compatibility of a model with DHARMa by trying to run various functions that are needed
}
\author{
Florian Hartig
}
