% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testOverdispersion.R
\name{testOverdispersion}
\alias{testOverdispersion}
\title{Simulated overdipersion test}
\usage{
testOverdispersion(simulationOutput, alternative = "overdispersion",
  print = F, plot = F)
}
\arguments{
\item{alternative}{whether to test for overdispersion, underdispersion, or both}

\item{model}{a fitted model object. See details for possible models}
}
\description{
Simulated overdipersion test
}
\details{
This test compares the approximate deviance (via squared pearson residuals) with the same quantity from a number of simulated models.
}
\seealso{
\code{\link{testSimulatedResiduals}}, \code{\link{testSimulatedResiduals}}, \code{\link{testZeroInflation}}, \code{\link{testTemporalAutocorrelation}}, \code{\link{testSpatialAutocorrelation}}
}

