% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testsResiduals.R
\name{testSimulatedResiduals}
\alias{testSimulatedResiduals}
\title{Wrapper for the various test functions}
\usage{
testSimulatedResiduals(simulationOutput)
}
\arguments{
\item{simulationOutput}{an object with simulated residuals created by \code{\link{simulateResiduals}}}
}
\description{
Wrapper for the various test functions
}
\details{
Currently, this function calls only the \code{\link{testUniformity}} function. All other tests (see below) have to be called by hand.
}
\note{
for special problems, in particular overdispersion, a \code{\link{parametricDispersionTest}} will likely be more powerful
}
\seealso{
\code{\link{testUniformity}}, \code{\link{testZeroInflation}}, \code{\link{testTemporalAutocorrelation}}, \code{\link{testSpatialAutocorrelation}}, \code{\link{testOverdispersion}}
}

