% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesisTestsResiduals.R
\name{testSpatialAutocorrelation}
\alias{testSpatialAutocorrelation}
\title{Convenience Function to test for spatial autocorrelation}
\usage{
testSpatialAutocorrelation(simulationOutput, x, y, print = T, plot = T)
}
\arguments{
\item{simulationOutput}{the result of a residual simulation}

\item{x}{the x coordinate, in the same order as the data points}

\item{y}{the x coordinate, in the same order as the data points}
}
\description{
Convenience Function to test for spatial autocorrelation
}
\details{
performs the Durbin-Watson Test against euklidian distance and plots the residuals against time
}
\note{
This function is provide as a base recommendation.
}

