% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{testSummary}
\alias{testSummary}
\title{Generic simulation test of a summary statistic}
\usage{
testSummary(simulationOutput, summary, alternative = c("greater", "two.sided",
  "less"), plot = T, methodName = "DHARMa generic simulation test")
}
\description{
Generic simulation test of a summary statistic
}
\author{
Florian Hartig
}
