% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testsResiduals.R
\name{testTemporalAutocorrelation}
\alias{testTemporalAutocorrelation}
\title{Test for temporal autocorrelation}
\usage{
testTemporalAutocorrelation(simulationOutput, time, print = T, plot = T)
}
\arguments{
\item{simulationOutput}{an object with simulated residuals created by \code{\link{simulateResiduals}}}

\item{time}{the time, in the same order as the data points}

\item{print}{whether to print output}

\item{plot}{whether to plot output}
}
\description{
Test for temporal autocorrelation
}
\details{
performs the Durbin-Watson Test and plots the residuals against time
}
\note{
This function is provide as a base recommendation.
}
\seealso{
\code{\link{testUniformDistribution}}, \code{\link{testZeroInflation}}, \code{\link{testSimulatedResiduals}}, \code{\link{testSpatialAutocorrelation}}
}

