% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{testUniformity}
\alias{testUniformity}
\title{Test for overall uniformity}
\usage{
testUniformity(simulationOutput)
}
\arguments{
\item{simulationOutput}{a DHARMa object with simulated residuals created with \code{\link{simulateResiduals}}}
}
\description{
This function tests the overall uniformity of the simulated residuals in a DHARMa object
}
\details{
The function applies a KS test for uniformity on the simulated residuals
}
\seealso{
\code{\link{testSimulatedResiduals}}, \code{\link{testZeroInflation}}, \code{\link{testTemporalAutocorrelation}}, \code{\link{testSpatialAutocorrelation}}, \code{\link{testOverdispersion}}, \code{\link{testOverdispersionParametric}}
}
