% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testResiduals.R
\name{testUnivariateDistribution}
\alias{testUnivariateDistribution}
\title{Tests residuals against a uniform distribution}
\usage{
testUnivariateDistribution(residuals, print = T)
}
\description{
Tests residuals against a uniform distribution
}
\details{
Tests against uniform distribution with KS test
}

