% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{testZeroInflation}
\alias{testZeroInflation}
\title{Tests for zero-inflation}
\usage{
testZeroInflation(simulationOutput, ...)
}
\arguments{
\item{simulationOutput}{a DHARMa object with simulated residuals created with \code{\link{simulateResiduals}}}

\item{...}{further arguments to \code{\link{testGeneric}}}
}
\description{
This function compares the observed number of zeros with the zeros expected from simulations.
}
\details{
shows the expected distribution of zeros against the observed
}
\seealso{
\code{\link{testResiduals}}, \code{\link{testUniformity}}, \code{\link{testDispersion}}, \code{\link{testGeneric}}, \code{\link{testTemporalAutocorrelation}}, \code{\link{testSpatialAutocorrelation}}
}
\author{
Florian Hartig
}
