% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformQuantiles.R
\name{transformQuantiles}
\alias{transformQuantiles}
\title{Transform quantiles to pdf (deprecated)}
\usage{
transformQuantiles(res, quantileFunction = qnorm, outlierValue = 7)
}
\description{
The purpose of this function was to transform the DHARMa quantile residuals (which have a uniform distribution) to a particular pdf. Since DHARMa 0.3.0, this functionality is integrated in the \code{\link{residuals.DHARMa}} function. Please switch to using this function.
}
