% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarkingTypeI-II.R
\name{benchmarkOverdispersion}
\alias{benchmarkOverdispersion}
\title{Simulated p-values for internal testing}
\usage{
benchmarkOverdispersion(dispersionValues = 0, n = 10, alpha = 0.05, ...)
}
\arguments{
\item{n}{replicates}

\item{alpha}{significance level}

\item{overdispersion}{amount of overdispersion}
}
\value{
list of various named objects. Within each, order is dispersion_glmer, overdisp_fun, DHARMa
}
\description{
Simulated p-values for internal testing
}

