% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateResiduals.R
\name{checkModel}
\alias{checkModel}
\title{Check if models is supported}
\usage{
checkModel(fittedModel, stop = F)
}
\arguments{
\item{fittedModel}{a model}

\item{stop}{whether to stop}
}
\description{
Check if models is supported
}
\keyword{internal}
