% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compatibility.R
\name{checkModel}
\alias{checkModel}
\title{Check if the fitted model is supported by DHARMa}
\usage{
checkModel(fittedModel, stop = F)
}
\arguments{
\item{fittedModel}{a fitted model.}

\item{stop}{whether to throw an error if the model is not supported by DHARMa.}
}
\description{
The function checks if the fitted model is supported by DHARMa, and if there are other issues that could create problems.
}
\details{
The main purpose of this function is to check if the fitted model class is supported by DHARMa. The function additionally checks for properties of the fitted model that could create problems for calculating residuals or working with the results in DHARMa.
}
\keyword{internal}
