% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createData.R
\name{createData}
\alias{createData}
\title{Function to simulate data}
\usage{
createData(replicates = 1, sampleSize = 10, intercept = 0,
  fixedEffects = 1, numGroups = 10, randomEffectVariance = 1,
  overdispersion = 0.5, family = poisson(), scale = 1, cor = 0)
}
\description{
Function to simulate data
}

