% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DHARMa.R
\docType{package}
\name{DHARMa}
\alias{DHARMa}
\alias{DHARMa-package}
\title{DHARMa}
\description{
Residual Diagnostics for HierArchical (Multi-level / Mixed) Regression Models
}
\details{
See index / vignette for details
}
\examples{
vignette("Demo", package="DHARMa")
}

