% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DHARMa.R
\docType{package}
\name{DHARMa}
\alias{DHARMa}
\alias{DHARMa-package}
\title{DHARMa - Residual Diagnostics for HierArchical (Multi-level / Mixed) Regression Models}
\description{
The DHARMa package uses a simulation-based approach to create readily interpretable scaled residuals from fitted generalized linear mixed models. Currently supported are lme4, glm (except quasi-distributions) and lm model classes. The resulting residuals are standardized to values between 0 and 1 and can be interpreted as intuitively as residuals from a linear regression. The package also provides a number of plot and test functions for typical model mispecification problem, such as over/underdispersion, zero-inflation, and spatial and temporal autocorrelation.
}
\details{
See index / vignette for details
}
\examples{
vignette("Demo", package="DHARMa")
}
\seealso{
\code{\link{simulatedResiduals}}
}

