% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{ensurePredictor}
\alias{ensurePredictor}
\title{Ensures the existence of a valid predictor to plot residuals against}
\usage{
ensurePredictor(simulationOutput, predictor = NULL)
}
\arguments{
\item{simulationOutput}{a DHARMa simulation output or an object that can be converted into a DHARMa simulation output}

\item{predictor}{an optional predictor. If no predictor is provided, will try to extract the fitted value}
}
\description{
Ensures the existence of a valid predictor to plot residuals against
}
\keyword{internal}
