% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compatibility.R
\name{getData}
\alias{getData}
\alias{getData.default}
\alias{getData.merMod}
\alias{getData.brmsfit}
\title{Get model data}
\usage{
getData(object, ...)

\method{getData}{default}(object, ...)

\method{getData}{merMod}(object, ...)

\method{getData}{brmsfit}(object, ...)
}
\arguments{
\item{object}{a fitted model.}

\item{...}{additional parameters to be passed on.}
}
\description{
Wrapper to get the data that was used to fit a model.
}
\note{
The data is retrieved from the environment where the model was fit, which is usually your current environment. If you delete the data from the environment or change it after fitting the model this function will probably not work or get the wrong dataset, respectively.
}
\seealso{
\link{getObservedResponse}, \link{getSimulations}, \link{getRefit}, \link{getFixedEffects}, \link{getFitted}
}
\author{
Florian Hartig
}
