% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compatibility.R
\name{getData}
\alias{getData}
\alias{getData.default}
\alias{getData.merMod}
\title{Get model data}
\usage{
getData(object, ...)

\method{getData}{default}(object, ...)

\method{getData}{merMod}(object, ...)
}
\arguments{
\item{object}{a fitted model.}

\item{...}{additional parameters to be passed on.}
}
\description{
Wrapper to get the data that was used to fit a model.
}
\seealso{
\link{getObservedResponse}, \link{getSimulations}, \link{getRefit}, \link{getFixedEffects}, \link{getFitted}
}
\author{
Florian Hartig
}
