% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compatibility.R
\name{getFamily}
\alias{getFamily}
\alias{getFamily.default}
\alias{getFamily.phylolm}
\alias{getFamily.phyloglm}
\title{Get model family}
\usage{
getFamily(object, ...)

\method{getFamily}{default}(object, ...)

\method{getFamily}{phylolm}(object, ...)

\method{getFamily}{phyloglm}(object, ...)
}
\arguments{
\item{object}{A fitted model.}

\item{...}{Additional parameters to be passed on.}
}
\description{
Wrapper to get the family of a fitted model.
}
\seealso{
\link{getObservedResponse}, \link{getSimulations}, \link{getRefit}, \link{getFixedEffects}, \link{getFitted}
}
\author{
Florian Hartig
}
