% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compatibility.R
\name{getFitted}
\alias{getFitted}
\title{Get model fitted}
\usage{
getFitted(object, ...)
}
\arguments{
\item{object}{a fitted model}

\item{...}{additional parameters to be passed on, usually to the simulate function of the respective model class}
}
\description{
Wrapper to get the fitted value a fitted model
}
\details{
The purpose of this wrapper is to standardize extract the fitted values
}
\author{
Florian Hartig
}
