% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compatibility.R
\name{getObservedResponse}
\alias{getObservedResponse}
\title{Get model response}
\usage{
getObservedResponse(object, ...)
}
\arguments{
\item{object}{a fitted model}

\item{...}{additional parameters}
}
\description{
Extract the response of a fitted model
}
\details{
The purpose of this function is to savely extract the response (dependent variable) of the fitted model classes
}
\author{
Florian Hartig
}
