% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compatibility.R
\name{getRefit}
\alias{getRefit}
\title{Get model refit}
\usage{
getRefit(object, newresp, ...)
}
\arguments{
\item{object}{a fitted model}

\item{...}{additional parameters to be passed on, usually to the simulate function of the respective model class}
}
\description{
Wrapper to refit a fitted model
}
\details{
The purpose of this wrapper is to standardize the refit of a model
}
\author{
Florian Hartig
}
