% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compatibility.R
\name{getResiduals}
\alias{getResiduals}
\alias{getResiduals.default}
\alias{getResiduals.MixMod}
\title{Get model residuals}
\usage{
getResiduals(object, ...)

\method{getResiduals}{default}(object, type = "response", ...)

\method{getResiduals}{MixMod}(object, ...)
}
\arguments{
\item{object}{a fitted model}

\item{...}{additional parameters}
}
\description{
Extract the residuals of a fitted model
}
\details{
The purpose of this function is to savely extract the residuals of the fitted model classes
}
\author{
Florian Hartig
}
