% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compatibility.R
\name{getSimulations}
\alias{getSimulations}
\title{Get model simulations}
\usage{
getSimulations(object, ...)
}
\description{
Simulate from a fitted model
}
\details{
The purpose of this function is to standardize the simulations from a model in a standardized way
}
\author{
Florian Hartig
}
