% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateResiduals.R
\name{multiModAssess}
\alias{multiModAssess}
\title{Apply DHARMa residual diagnostic tests to a candidate set of models.}
\usage{
multiModAssess(modList, n = 500)
}
\arguments{
\item{modList}{A list of model objects.}

\item{n}{Number of DHARMa simulations to generate from each model.}
}
\value{
Returns a data.frame containing the following columns:

\code{statistic}: Test statistics for each particular test.

\code{p.value}: P values from each test.

\code{alt.hyp}: Alternative hypothesis specified in each test.

\code{method}: A short description of each applied test.

\code{response}: LHS of specified model formula.

\code{predictor}: RHS of specified model formula.
}
\description{
This function applies DHARMa residual diagnostic tests to a user-specified list of models. Tests
look for agreement between simulated and empirical residuals; assessing models for dispersion,
zero-inflation, anomalous prevalence of outliers, and deviations from expected distributions.
}
\details{
This functionality emerged from user interest in applying DHARMa residual diagnostic tests to several
models in a model selection context. Model selection necessitates that candidate models meet
modeling assumptions regardless of complexity (Mazerolle 2019). Otherwise, the most "appropriate" model
may be parsimonious but misspecified.

This function applies the tests within the \code{testResdiuals} function, as well as
\code{testZeroInflation}. Caution: tests may not be applicable in every context.
}
\examples{

library(glmmTMB)
modList <- list()
modList[[1]] <- glmmTMB(count ~ mined + (1|site), zi=~mined, family=poisson, data=Salamanders)
modList[[2]] <- glmmTMB(count ~ mined + cover + (1|site), zi=~mined, family=poisson, data=Salamanders)
modList[[3]] <- glmmTMB(count ~ mined + cover + spp + (1|site), zi=~mined, family=poisson, data=Salamanders)
multiModAssess(modList)
}
\references{
Mazerolle, Marc J. 2019. AICcmodavg: Model Selection and Multimodel Inference Based on (Q)AIC(c). https://cran.r-project.org/package=AICcmodavg.
}
