% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DHARMa.R
\name{plot.DHARMa}
\alias{plot.DHARMa}
\title{Plot simulated residuals}
\usage{
\method{plot}{DHARMa}(x, ...)
}
\arguments{
\item{x}{an object with simulated residuals created by \code{\link{simulateResiduals}}}

\item{...}{optional arguments to pass on to \code{\link{plotSimulatedResiduals}}}
}
\description{
Plot simulated residuals
}
\details{
The function calls \code{\link{plotSimulatedResiduals}} - see details there
}
