% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runBenchmarks.R
\name{plot.DHARMaBenchmark}
\alias{plot.DHARMaBenchmark}
\title{Plots DHARMa benchmark}
\usage{
\method{plot}{DHARMaBenchmark}(x, ...)
}
\arguments{
\item{x}{object of class DHARMaBenchmark, created by \code{\link{runBenchmarks}}}

\item{...}{parameters to pass to the plot function}
}
\description{
Plots DHARMa benchmark
}
