% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSimulatedResiduals.R
\name{plotConventionalResiduals}
\alias{plotConventionalResiduals}
\title{Convenience function to draw conventional residual plots}
\usage{
plotConventionalResiduals(fittedModel)
}
\arguments{
\item{fittedModel}{a fitted model object}
}
\description{
Convenience function to draw conventional residual plots
}

