% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSimulatedResiduals.R
\name{plotResiduals}
\alias{plotResiduals}
\title{Plots a generic residual plot with spline}
\usage{
plotResiduals(pred, res, quantreg = T)
}
\description{
Plots a generic residual plot with spline
}

