% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotResiduals.R
\name{plotResiduals}
\alias{plotResiduals}
\title{Generic residual plot with either spline or quantile regression}
\usage{
plotResiduals(pred, residual, quantreg = T, ...)
}
\arguments{
\item{pred}{predictor variable}

\item{residual}{residual variable}

\item{quantreg}{should a quantile regression be performed. If F, a smooth spline will be}

\item{...}{additional arguments to plot}
}
\description{
Generic residual plot with either spline or quantile regression
}
\seealso{
\code{\link{plotSimulatedResiduals}}
}

