% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotSimulatedResiduals}
\alias{plotSimulatedResiduals}
\title{DHARMa standard residual plots}
\usage{
plotSimulatedResiduals(simulationOutput, ...)
}
\arguments{
\item{simulationOutput}{an object with simulated residuals created by \link{simulateResiduals}}

\item{...}{further options for \link{plotResiduals}. Consider in particular parameters quantreg, rank and asFactor. xlab, ylab and main cannot be changed when using plotSimulatedResiduals, but can be changed when using plotResiduals.}
}
\description{
DEPRECATED, use plot() instead
}
\note{
This function is deprecated. Use \link{plot.DHARMa}
}
\seealso{
\link{plotResiduals}, \link{plotQQunif}
}
