% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSimulatedResiduals.R
\name{plotSimulatedResiduals}
\alias{plotSimulatedResiduals}
\title{creates Poisson data overdispersion and random intercept}
\usage{
plotSimulatedResiduals(simulationOutput, quantreg = T)
}
\arguments{
\item{simulationOutput}{a result from simulatedResiduals}

\item{quantreg}{whether to perform a quantile regression on 0.25, 0.5, 0.75}
}
\description{
creates Poisson data overdispersion and random intercept
}

