% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DHARMa.R
\name{print.DHARMa}
\alias{print.DHARMa}
\title{Print simulated residuals}
\usage{
\method{print}{DHARMa}(x, ...)
}
\arguments{
\item{x}{an object with simulated residuals created by \code{\link{simulateResiduals}}}

\item{...}{optional arguments to pass on to \code{\link{plotSimulatedResiduals}}}
}
\description{
Print simulated residuals
}

