% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DHARMa.R
\name{print.DHARMa}
\alias{print.DHARMa}
\title{Print simulated residuals}
\usage{
\method{print}{DHARMa}(simulationOutput)
}
\arguments{
\item{simulationOutput}{an object with simulated residuals created by \code{\link{simulateResiduals}}}
}
\description{
Print simulated residuals
}
