% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz-testTypeI.R
\name{simulatePvaluesDispersion}
\alias{simulatePvaluesDispersion}
\title{Simulated p-values for internal testing}
\usage{
simulatePvaluesDispersion(overdispersion = 0, n = 20, alpha = 0.05)
}
\value{
list of various named objects. Within each, order is dispersion_glmer, overdisp_fun, DHARMa
}
\description{
Simulated p-values for internal testing
}

