% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateResiduals.R
\name{simulateResiduals}
\alias{simulateResiduals}
\title{Creates scaled residuals by simulation}
\usage{
simulateResiduals(fittedModel, n = 250, refit = F, integerResponse = NULL)
}
\arguments{
\item{fittedModel}{fitted model object, currently restricted to lme4, lm, or glm models}

\item{n}{integer number > 1, number of simulations to run. Set at least 250, better 1000}

\item{refit}{T or F - should the model be refit to do a parametric bootstrap}

\item{integerResponse}{T or F - is this a model with an integer response (such as Poisson or Binomial). If not provided, the function will attept to find out by itself, may not work for all families / distributions}
}
\description{
Creates scaled residuals by simulation
}
\details{
The integerResponse option essentially adds a uniform noise from -0.5 to 0.5 on the simulated and observed response. Note that this works because the expected distribution of this is flat - you can see this via hist(ecdf(runif(10000))(runif(10000)))
}

