% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateResiduals.R
\name{testApply}
\alias{testApply}
\title{Apply tests to model and extract summary statistics}
\usage{
testApply(mod, n = 500)
}
\arguments{
\item{mod}{A model}

\item{n}{Number of DHARMa simulations}
}
\description{
This function applies a set of DHARMa residual tests to a single model
}
\keyword{internal}
