% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compatibility.R
\name{testModel}
\alias{testModel}
\title{Test DHARMa compatibility}
\usage{
testModel(fittedModel)
}
\value{
A report of possible problems
}
\description{
The function tests the compatibility of a model with DHARMa
}
\author{
Florian Hartig
}
