% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testsResiduals.R
\name{testOverdispersion}
\alias{testOverdispersion}
\title{Test for over/underdispersion}
\usage{
testOverdispersion(simulationOutput, alternative = "overdispersion",
  plot = F)
}
\arguments{
\item{simulationOutput}{an object with simulated residuals created by \code{\link{simulateResiduals}}}

\item{alternative}{whether to test for "overdispersion", "underdispersion", or "both" (both reduces power)}

\item{plot}{whether to plot output}
}
\description{
This function performs a simulation-based test for over/underdispersion
}
\details{
The function implements two tests, depending on whether it is applied on a simulation with refit = F, or refit = T. 

If refit = F (default, and recommended), the function tests if the IQR of the scaled residuals deviate from the null hypothesis of a uniform distribution. 

If refit = T, the function compares the approximate deviance (via squared pearson residuals) with the same quantity from the models refit with simulated models. It is MUCH slower than the option above and the parametric alternative \code{\link{testOverdispersionParametric}}. At the moment it's only implemented for testing purposes.
}
\seealso{
\code{\link{testSimulatedResiduals}}, \code{\link{testSimulatedResiduals}}, \code{\link{testZeroInflation}}, \code{\link{testTemporalAutocorrelation}}, \code{\link{testSpatialAutocorrelation}}, \code{\link{testOverdispersionParametric}}
}
