% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{testOverdispersion}
\alias{testOverdispersion}
\title{Simulated overdisperstion tests}
\usage{
testOverdispersion(simulationOutput, ...)
}
\arguments{
\item{simulationOutput}{an object of class DHARMa with simulated quantile residuals, either created via \code{\link{simulateResiduals}} or by \code{\link{createDHARMa}} for simulations created outside DHARMa}

\item{...}{additional arguments to \code{\link{testDispersion}}}
}
\description{
Simulated overdisperstion tests
}
\details{
Deprecated, switch your code to using the \code{\link{testDispersion}} function
}
