% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{testOverdispersion}
\alias{testOverdispersion}
\title{Simulated overdisperstion tests (deprecated)}
\usage{
testOverdispersion(simulationOutput, ...)
}
\arguments{
\item{simulationOutput}{an object of class DHARMa with simulated quantile residuals, either created via \link{simulateResiduals} or by \link{createDHARMa} for simulations created outside DHARMa}

\item{...}{additional arguments to \link{testDispersion}}
}
\description{
Simulated overdisperstion tests (deprecated)
}
\details{
Deprecated, switch your code to using the \link{testDispersion} function
}
