% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{testOverdispersionParametric}
\alias{testOverdispersionParametric}
\title{Parametric overdisperstion tests}
\usage{
testOverdispersionParametric(...)
}
\arguments{
\item{...}{arguments will be ignored, the parametric tests is no longer recommend}
}
\description{
Parametric overdisperstion tests
}
\details{
Deprecated, switch your code to using the \code{\link{testDispersion}} function. The function will do nothing, arguments will be ignored, the parametric tests is no longer recommend
}
