% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarkPvalues.R
\name{testPDistribution}
\alias{testPDistribution}
\title{Plot distribution of p-values}
\usage{
testPDistribution(x, plot = T, title = "p value", ...)
}
\arguments{
\item{x}{vector of p values}

\item{plot}{should the values be plottet}

\item{...}{additional arguments to hist}
}
\description{
Plot distribution of p-values
}
\author{
Florian Hartig
}
