% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runBenchmarks.R
\name{testPDistribution}
\alias{testPDistribution}
\title{Plot distribution of p-values.}
\usage{
testPDistribution(x, plot = TRUE,
  main = "p distribution \\n expected is flat at 1", ...)
}
\arguments{
\item{x}{Vector of p values.}

\item{plot}{Should the values be plotted.}

\item{main}{Title for the plot.}

\item{...}{Additional arguments to hist.}
}
\description{
Plot distribution of p-values.
}
\author{
Florian Hartig
}
