% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{testQuantiles}
\alias{testQuantiles}
\title{Test for quantiles}
\usage{
testQuantiles(simulationOutput, predictor = NULL, quantiles = c(0.25,
  0.5, 0.75), plot = T)
}
\arguments{
\item{simulationOutput}{an object of class DHARMa with simulated quantile residuals, either created via \code{\link{simulateResiduals}} or by \code{\link{createDHARMa}} for simulations created outside DHARMa}

\item{predictor}{an optional predictor variable to be used, instead of the predicted response (default)}

\item{quantiles}{the quantiles to be tested}

\item{plot}{if T, the function will create an additional plot}
}
\description{
This function tests
}
\details{
The function fits quantile gams for the specified quantiles on the residuals, and reports the p-values for each individually, as well as a combined p-value that is adjusted for multiple testing.
}
\seealso{
\code{\link{testResiduals}}, \code{\link{testUniformity}}, \code{\link{testDispersion}}, \code{\link{testZeroInflation}}, \code{\link{testGeneric}}, \code{\link{testTemporalAutocorrelation}}, \code{\link{testSpatialAutocorrelation}}
}
\author{
Florian Hartig
}
