% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{testResiduals}
\alias{testResiduals}
\title{DHARMa residual tests}
\usage{
testResiduals(simulationOutput, type = c("uniformity", "dispersion",
  "summary"), ...)
}
\arguments{
\item{simulationOutput}{a DHARMa object with simulated residuals created with \code{\link{simulateResiduals}}}

\item{type}{a string or a vector of string with the tests to be performed.}

\item{...}{parameters to be passed on to the test.}
}
\description{
Wrapper for all residual tests implemented in DHARMa
}
\details{
This function is a wrapper for the various test functions implemented in DHARNa. Currently, this function calls only the \code{\link{testUniformity}} function. Other tests (see below) have to be called by hand

Currently, this function calls only the \code{\link{testUniformity}} function. All other tests (see below) have to be called by hand.
}
\seealso{
\code{\link{testUniformity}}, \code{\link{testZeroInflation}}, \code{\link{testTemporalAutocorrelation}},
}
