% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{testResiduals}
\alias{testResiduals}
\title{DHARMa general residual test}
\usage{
testResiduals(simulationOutput)
}
\arguments{
\item{simulationOutput}{a DHARMa object with simulated residuals created with \code{\link{simulateResiduals}}}
}
\description{
Calls both uniformity and dispersion test
}
\details{
This function is a wrapper for the various test functions implemented in DHARMa. Currently, this function calls the \code{\link{testUniformity}} and the \code{\link{testDispersion}} functions. All other tests (see below) have to be called by hand.
}
\seealso{
\code{\link{testUniformity}}, \code{\link{testDispersion}}, \code{\link{testZeroInflation}}, \code{\link{testGeneric}}, \code{\link{testTemporalAutocorrelation}}, \code{\link{testSpatialAutocorrelation}}
}
\author{
Florian Hartig
}
