% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesisTestsResiduals.R
\name{testSimulatedResiduals}
\alias{testSimulatedResiduals}
\title{Performs automatic tests of the simulated residuals}
\usage{
testSimulatedResiduals(simulationOutput)
}
\arguments{
\item{simulationOutput}{a simulation output}
}
\description{
Performs automatic tests of the simulated residuals
}

