% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{testSimulatedResiduals}
\alias{testSimulatedResiduals}
\title{Residual tests}
\usage{
testSimulatedResiduals(simulationOutput)
}
\arguments{
\item{simulationOutput}{a DHARMa object with simulated residuals created with \code{\link{simulateResiduals}}}
}
\description{
Residual tests
}
\details{
Deprecated, switch your code to using the \code{\link{testResiduals}} function
}
\author{
Florian Hartig
}
