% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{testSimulatedResiduals}
\alias{testSimulatedResiduals}
\title{Residual tests}
\usage{
testSimulatedResiduals(simulationOutput)
}
\arguments{
\item{simulationOutput}{an object of class DHARMa, either created via \code{\link{simulateResiduals}} for supported models or by \code{\link{createDHARMa}} for simulations created outside DHARMa, or a supported model. Providing a supported model directly is discouraged, because simulation settings cannot be changed in this case.}
}
\description{
Residual tests
}
\details{
Deprecated, switch your code to using the \code{\link{testResiduals}} function
}
\author{
Florian Hartig
}
