% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testsResiduals.R
\name{testSpatialAutocorrelation}
\alias{testSpatialAutocorrelation}
\title{Convenience Function to test for spatial autocorrelation}
\usage{
testSpatialAutocorrelation(simulationOutput, x, y, print = T, plot = T)
}
\arguments{
\item{simulationOutput}{an object with simulated residuals created by \code{\link{simulateResiduals}}}

\item{x}{the x coordinate, in the same order as the data points}

\item{y}{the x coordinate, in the same order as the data points}

\item{print}{whether to print output}

\item{plot}{whether to plot output}
}
\description{
Convenience Function to test for spatial autocorrelation
}
\details{
performs the Durbin-Watson Test against euklidian distance and plots the residuals against time
}
\note{
This function is provide as a base recommendation.
}
\seealso{
\code{\link{testUniformDistribution}}, \code{\link{testZeroInflation}}, \code{\link{testTemporalAutocorrelation}}, \code{\link{testSimulatedResiduals}}
}

