% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesisTestsResiduals.R
\name{testTemporalAutocorrelation}
\alias{testTemporalAutocorrelation}
\title{Test for temporal autocorrelation}
\usage{
testTemporalAutocorrelation(simulationOutput, time, print = T, plot = T)
}
\arguments{
\item{simulationOutput}{the result of a residual simulation}

\item{time}{the time, in the same order as the data points}
}
\description{
Test for temporal autocorrelation
}
\details{
performs the Durbin-Watson Test and plots the residuals against time
}
\note{
This function is provide as a base recommendation.
}

