% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testsResiduals.R
\name{testUniformity}
\alias{testUniformity}
\title{Tests residuals against a uniform distribution}
\usage{
testUniformity(simulationOutput, print = T)
}
\arguments{
\item{simulationOutput}{an object with simulated residuals created by \code{\link{simulateResiduals}}}

\item{print}{whether to print output}
}
\description{
Tests residuals against a uniform distribution
}
\details{
Tests residuals against a uniform distribution with the KS test
}
\seealso{
\code{\link{testSimulatedResiduals}}, \code{\link{testZeroInflation}}, \code{\link{testTemporalAutocorrelation}}, \code{\link{testSpatialAutocorrelation}}, \code{\link{testOverdispersion}}
}

