% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testsResiduals.R
\name{testUniformity}
\alias{testUniformity}
\title{Test for overall uniformity}
\usage{
testUniformity(simulationOutput)
}
\arguments{
\item{simulationOutput}{an object with simulated residuals created by \code{\link{simulateResiduals}}}
}
\description{
This function tests the overall uniformity of the residuals
}
\details{
Tests residuals against a uniform distribution with the KS test
}
\seealso{
\code{\link{testSimulatedResiduals}}, \code{\link{testZeroInflation}}, \code{\link{testTemporalAutocorrelation}}, \code{\link{testSpatialAutocorrelation}}, \code{\link{testOverdispersion}}, \code{\link{testOverdispersionParametric}}
}
