% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testsResiduals.R
\name{testZeroInflation}
\alias{testZeroInflation}
\title{Tests for zero-inflation}
\usage{
testZeroInflation(simulationOutput, plot = T, alternative = "more")
}
\arguments{
\item{simulationOutput}{a DHARMa object with simulated residuals created with \code{\link{simulateResiduals}}}

\item{plot}{whether to plot output}

\item{alternative}{whether to test for 'more', 'less', or 'both' more or less zeros in the observed data}
}
\description{
This function compares the observed number of zeros with the zeros expected from simulations.
}
\details{
shows the expected distribution of zeros against the observed
}
\seealso{
\code{\link{testUniformity}}, \code{\link{testSimulatedResiduals}}, \code{\link{testTemporalAutocorrelation}}, \code{\link{testSpatialAutocorrelation}}, \code{\link{testOverdispersion}}, \code{\link{testOverdispersionParametric}}
}
