% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesisTestsResiduals.R
\name{testZeroInflation}
\alias{testZeroInflation}
\title{Tests for zero-inflation in the residuals}
\usage{
testZeroInflation(simulationOutput, print = T, plot = T)
}
\description{
Tests for zero-inflation in the residuals
}
\details{
Tests against uniform distribution with KS test
}

