% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformQuantiles.R
\name{transformQuantiles}
\alias{transformQuantiles}
\title{Transform quantiles to pdf (deprecated)}
\usage{
transformQuantiles(res, quantileFunction = qnorm, outlierValue = 7)
}
\arguments{
\item{res}{an object with simulated residuals created by \link{simulateResiduals}}

\item{quantileFunction}{optional - a quantile function to transform the uniform 0/1 scaling of DHARMa to another distribution}

\item{outlierValue}{if a quantile function with infinite support (such as dnorm) is used, residuals that are 0/1 are mapped to -Inf / Inf. outlierValues allows to convert -Inf / Inf values to an optional min / max value.}
}
\description{
The purpose of this function was to transform the DHARMa quantile residuals (which have a uniform distribution) to a particular pdf. Since DHARMa 0.3.0, this functionality is integrated in the \link{residuals.DHARMa} function. Please switch to using this function.
}
