% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformQuantiles.R
\name{transformQuantiles}
\alias{transformQuantiles}
\title{Transform quantiles to pdf}
\usage{
transformQuantiles(res, quantileFunction = qnorm, outlierValue = 7)
}
\arguments{
\item{res}{DHARMa residuals}

\item{quantileFunction}{the quantile function of the desired distribution}

\item{outlierValue}{the value that should be assigned to residuals that are 0/1 and thus typically mapped to -Inf / Inf for continous distributions with infinte support}
}
\description{
The purpose of this function is to transform the DHARMa quantile residuals (which have a uniform distribution) to a particular pdf.
}
\details{
See also https://github.com/florianhartig/DHARMa/issues/39
}
\examples{
plot(res$fittedPredictedResponse, transformQuantiles(res))
}
